Title query has been extracted into the component
